+++ atomic_reactor/utils/operator.py
